Add files via upload
In this tutorial, we'll guide beginners through the process of creating a calculator using HTML, CSS, and JavaScript. We'll start by designing a user-friendly interface using HTML and styling it with CSS. After that, we'll delve into the world of JavaScript to add the calculator's essential functionality. This project is perfect for JavaScript newcomers, providing a hands-on experience in building a practical application. Join us as we create a dynamic calculator that combines form and function, making it a great starting point for those looking to enhance their JavaScript skills.